+++ frontend/vite.config.ts
@@ -1,10 +1,9 @@
 import path from 'path';
 import react from '@vitejs/plugin-react';
-import { defineConfig, loadEnv } from 'vite';
+import { defineConfig } from 'vite';
 
 // https://vite.dev/config/
 export default defineConfig(({ mode }) => {
-  const env = loadEnv(mode, process.cwd(), '');
   return {
     base: '',
     plugins: [react()],